+++ pyproject.toml
@@ -16,3 +16,4 @@ pytest = "^3.4"
 pytest-flask = "*"
 pytest-sugar = "*"
 pytest-cov = "*"
+blinker = "^1.4"
